# Remove Vector{Any} from input data