@odow 2023-05-24
I'm about to go to bed, but I'm very interested to see your new plots. It's such a shame that Julia makes it easy to write fast code, and easy to write horrifically slow code, and it's very hard to tell the difference.