@odow 2023-05-24
I also didn't add my supply chain model:
```
function jump(I, L, M, IJ, JK, IK, KL, LM, D, solve)
    model = if solve == "True"
        direct_model(Gurobi.Optimizer())
    else
        Model()
    end
    set_string_names_on_creation(model, false)
    x_list = [
        (i, j, k)
        for (i, j) in IJ
        for (jj, k) in JK if jj == j
        for (ii, kk) in IK if ii == i && kk == k
    ]
    y_list = [
        (i, k, l)
        for i in I
        for (k, l) in KL
    ]
    z_list = [(i, l, m) for i in I for (l, m) in LM]
    @variable(model, x[x_list] >= 0)
    @variable(model, y[y_list] >= 0)
    @variable(model, z[z_list] >= 0)
    for (i, k) in IK
        @constraint(model,
            sum(
                x[(i, j, k)]
                for (ii, j) in IJ if ii == i
                for (jj, kk) in JK if jj == j && kk == k
            ) >= sum(y[(i, k, l)] for (kk, l) in KL if kk == k)
        )
    end
    for i in I, l in L
        @constraint(model,
            sum(y[(i, k, l)] for (k, ll) in KL if ll == l) >=
            sum(z[(i, l, m)] for (ll, m) in LM if ll == l))
    end
    for i in I, m in M
        @constraint(model, sum(z[(i, l, m)] for (l, mm) in LM if mm == m) >= D[i, m])
    end
    if solve == "True"
        set_silent(model)
        optimize!(model)
    end
end
```

It'd be interesting to see the improvement that makes. But it's not essential. Ideally, normal JuMP should be fast enough.