@justine18 2023-05-24
Thanks for your contribution @odow. It indeed was a game changer. I guess the data structure did most of the improvements.

Here are the plot for IJKLM and supply chain.
**IJKLM**
![model_performance](https://github.com/justine18/performance_experiment/assets/15107381/bb5ee344-ec7d-4020-8848-a3c7104c75b2)

**Supply Chain**
![model_performance](https://github.com/justine18/performance_experiment/assets/15107381/0d64c6f7-a958-473d-9322-567f734b0740)